requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.4
exceptiongroup==1.2.0
faiss-gpu==1.7.2
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
inflection==0.5.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.14
langchain-community==0.0.30
langchain-core==0.1.37
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.38
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.14.3
orjson==3.10.0
packaging==23.2
pathspec==0.12.1
pinecone-client==3.2.2
platformdirs==4.2.0
psycopg2==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.6.0
tomli==2.0.1
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
